requirements.txt
asgiref==3.2.10
Django==3.1.1
Pillow==2.2.1
pytesseract==0.3.6
pytz==2020.1
sqlparse==0.3.1
